Move enum to response class
